Better map loader

+++ Mario/Mario/Environment/Ground.cpp
@@ -40,5 +40,5 @@ void Ground::HandleCollisionStart(Collider* other)
 
 int Ground::CollidersVectorSize()
 {
-	return boxColliders.size();
+	return (int)boxColliders.size();
 }
